+++ admin/app.js
@@ -93,6 +93,7 @@ if (featureTogglesMerged) {
 const index = require('./routes/index')
 const testDeveloper = require('./routes/test-developer')
 const serviceManager = require('./routes/service-manager')
+const serviceMessage = require('./routes/service-message')
 const helpdesk = require('./routes/helpdesk')
 const school = require('./routes/school')
 const pupilPin = require('./routes/pupil-pin')
@@ -273,6 +274,7 @@ if (WEBSITE_OFFLINE) {
   app.use('/', index)
   app.use('/test-developer', testDeveloper)
   app.use('/service-manager', serviceManager)
+  app.use('/service-manager/service-message', serviceMessage)
   app.use('/helpdesk', helpdesk)
   app.use('/school', school)
   app.use('/pupil-pin', pupilPin)
