+++ admin/spec/back-end/service/administration-message.service.spec.js
@@ -47,7 +47,6 @@ describe('administrationMessageService', () => {
       spyOn(administrationMessageService, 'prepareSubmissionData')
         .and.returnValue({ title: 'serviceMessageTitle', message: 'serviceMessageContent' })
       spyOn(administrationMessageDataService, 'sqlCreate')
-      spyOn(administrationMessageDataService, 'sqlUpdate')
       spyOn(redisCacheService, 'set')
     })
     it('should not continue further if a user id is not present', async () => {
@@ -64,13 +63,12 @@ describe('administrationMessageService', () => {
       expect(emptyFieldsValidator.validate).not.toHaveBeenCalled()
       expect(administrationMessageService.prepareSubmissionData).not.toHaveBeenCalled()
       expect(administrationMessageDataService.sqlCreate).not.toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlUpdate).not.toHaveBeenCalled()
       expect(redisCacheService.set).not.toHaveBeenCalled()
     })
     it('should call emptyFieldsValidator.validate', async () => {
       const validationError = new ValidationError()
       spyOn(emptyFieldsValidator, 'validate').and.returnValues(validationError)
-      const userId = '1'
+      const userId = 1
       await administrationMessageService.setMessage(requestData, userId)
       expect(emptyFieldsValidator.validate).toHaveBeenCalled()
     })
@@ -78,68 +76,42 @@ describe('administrationMessageService', () => {
       const validationError = new ValidationError()
       validationError.addError('serviceMessageTitle', 'error')
       spyOn(emptyFieldsValidator, 'validate').and.returnValues(validationError)
-      const userId = '1'
+      const userId = 1
       await administrationMessageService.setMessage(requestData, userId)
       expect(emptyFieldsValidator.validate).toHaveBeenCalled()
       expect(administrationMessageService.prepareSubmissionData).not.toHaveBeenCalled()
       expect(administrationMessageDataService.sqlCreate).not.toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlUpdate).not.toHaveBeenCalled()
       expect(redisCacheService.set).not.toHaveBeenCalled()
     })
-    it('should call administrationMessageDataService.sqlCreate if not in edit mode', async () => {
+    it('should call administrationMessageDataService.sqlCreate', async () => {
       const validationError = new ValidationError()
       spyOn(emptyFieldsValidator, 'validate').and.returnValues(validationError)
-      const userId = '1'
-      requestData.isEditView = undefined
+      const userId = 1
       await administrationMessageService.setMessage(requestData, userId)
       expect(emptyFieldsValidator.validate).toHaveBeenCalled()
       expect(administrationMessageService.prepareSubmissionData).toHaveBeenCalled()
       expect(administrationMessageDataService.sqlCreate).toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlUpdate).not.toHaveBeenCalled()
     })
-    it('should call administrationMessageDataService.sqlUpdate if in edit mode', async () => {
+    it('should call redisCacheService.set after a successful database transmission', async () => {
       const validationError = new ValidationError()
       spyOn(emptyFieldsValidator, 'validate').and.returnValues(validationError)
-      const userId = '1'
-      requestData.isEditView = true
+      const userId = 1
       await administrationMessageService.setMessage(requestData, userId)
       expect(emptyFieldsValidator.validate).toHaveBeenCalled()
       expect(administrationMessageService.prepareSubmissionData).toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlCreate).not.toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlUpdate).toHaveBeenCalled()
-    })
-    it('should call edisCacheService.set after a successful database transmission', async () => {
-      const validationError = new ValidationError()
-      spyOn(emptyFieldsValidator, 'validate').and.returnValues(validationError)
-      const userId = '1'
-      requestData.isEditView = true
-      await administrationMessageService.setMessage(requestData, userId)
-      expect(emptyFieldsValidator.validate).toHaveBeenCalled()
-      expect(administrationMessageService.prepareSubmissionData).toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlCreate).not.toHaveBeenCalled()
-      expect(administrationMessageDataService.sqlUpdate).toHaveBeenCalled()
+      expect(administrationMessageDataService.sqlCreate).toHaveBeenCalled()
       expect(redisCacheService.set).toHaveBeenCalledWith(serviceMessageRedisKey, { title: 'serviceMessageTitle', message: 'serviceMessageContent' })
     })
   })
   describe('prepareSubmissionData', () => {
-    it('should return an object that includes title, message and recordedByUser_id when creating a record', () => {
+    it('should return an object that includes title, message and createdByUser_id when creating a record', () => {
       const requestData = {
         serviceMessageTitle: 'serviceMessageTitle',
         serviceMessageContent: 'serviceMessageContent'
       }
       const userId = 1
       const result = administrationMessageService.prepareSubmissionData(requestData, userId)
-      expect(result).toEqual({ recordedByUser_id: 1, title: 'serviceMessageTitle', message: 'serviceMessageContent' })
-    })
-    it('should return an object that includes title, message and updatedByUser_id when creating a record', () => {
-      const requestData = {
-        isEditView: true,
-        serviceMessageTitle: 'serviceMessageTitle',
-        serviceMessageContent: 'serviceMessageContent'
-      }
-      const userId = 1
-      const result = administrationMessageService.prepareSubmissionData(requestData, userId)
-      expect(result).toEqual({ updatedByUser_id: 1, title: 'serviceMessageTitle', message: 'serviceMessageContent' })
+      expect(result).toEqual({ createdByUser_id: 1, title: 'serviceMessageTitle', message: 'serviceMessageContent' })
     })
   })
   describe('dropMessage', () => {
