+++ admin/views/service-manager/service-manager-home.ejs
@@ -29,7 +29,7 @@
                 </li>
                 <li>
                     <h2 class="govuk-heading-s"><a class="govuk-link" href="/service-manager/service-message">Manage service message</a></h2>
-                    <p>Create, edit or remove a service message</p>
+                    <p>Add or delete a service message for schools</p>
                 </li>
             </ol>
         </div>
