+++ admin/views/service-manager/service-message/service-message-form.ejs
@@ -45,7 +45,7 @@
                            id="serviceMessageTitle"
                            class="govuk-input govuk-!-width-three-quarters <%= err.isError('serviceMessageTitle') ? 'govuk-input--error' : '' %>"
                            name="serviceMessageTitle"
-                           value="<%= formData['serviceMessageTitle'] || formData['title'] %>"/>
+                           value="<%= formData['title'] %>"/>
                     <div class="govuk-!-margin-top-5 govuk-form-group <%= err.isError('serviceMessageContent') ? 'govuk-form-group--error' : '' %>">
                         <div class="govuk-label govuk-!-font-weight-bold" for="serviceMessageContent">Message</div>
                         <% if (err.isError('serviceMessageContent')) { %>
@@ -55,12 +55,9 @@
                                 class="govuk-textarea govuk-!-width-three-quarters <%= err.isError('serviceMessageContent') ? 'govuk-textarea--error' : '' %>"
                                 name="serviceMessageContent"
                                 id="serviceMessageContent"
-                                rows="4"><%= formData['serviceMessageContent'] || formData['message'] %></textarea>
+                                rows="4"><%= formData['message'] %></textarea>
                     </div>
                 </div>
-                <% if(isEditView) { %>
-                    <input type="hidden" id="isEditView" name="isEditView" value="true"/>
-                <% } %>
                 <button class="govuk-button govuk-!-margin-right-2" id="submit-service-message-button" type="submit">Save</button>
                 <a href="/service-manager/service-message" class="govuk-button govuk-button--secondary">Cancel</a>
             </form>
